// Downloads/dpm-ui-web-and-mobile-fixes 1/dpm-ui-web-and-mobile-fixes/web-apps/consumer-portal/src/types/quoteAndBuy.ts
interface Benefit {
    benefitCategory: string;
    benefitCode: string;
    benefitId: number;
    benefitNameAr: string;
    benefitNameEn: string;
    benefitPrice: number;
    description: string;
    mostPurchased: number;
}

interface PremiumBreakdown {
    amount: number;
    percentage?: number | null;
    sign: number;
    type: string;
}

interface TaxFeeBreakdown {
    amount: number;
    percentage?: number;
    type: string;
}

interface PricingOption {
    deductibleReferenceNo: string;
    deductibleAmount: number;
    finalAmount: number;
    premiumBreakdowns: PremiumBreakdown[];
    premiumDue: number;
    taxFeeBreakdowns: TaxFeeBreakdown[];
}

interface Vehicle {
    sequenceNo: string;
    vehicleMakeEnglishArabic: string;
    vehicleModelEnglishArabic: string;
}

interface PolicyDetails {
    benefits: Benefit[];
    policyEffectiveDate: string;
    policyExpiryDate: string;
    previousPolicyNo: string;
    pricingOptions: PricingOption[];
    quotationDate: string;
    quotationExpiryDate: string;
    renewalInd: string;
    vehicles: Vehicle[];
}